downloading updates
Update to save after each file, preventing massive number of retries if downloading a huge number of podcasts and it fails.
Update to skip files that already exist, useful when using in addition to other software.
Update to make sure if a file fails to download and a chunk was already downloaded, that we remove that chunk.