Add prune command
Add ability to prune old episodes via prune command.  Max age in days
can be set via configuration or passed in via command line.  This
requires tracking of the filename used to store the podcast episode,
so this has been added to the feed json.  Pulling the filename out of
the url can be used as a fallback; however, if the name has been pulled
out of the HTTP headers, then this feature will not be backwards
compatible for episodes named in this way and thus those will not be
pruned.